+++ example/src/main/scala/example/stockMarket/wealthManagement.scala
@@ -39,7 +39,7 @@ class WealthManagement(var initWealth: Double, val interestRate: Double) {
         cash += shares * dividendPerShare
     }
 
-    def takeAction(stockPrice: Double, marketState: List[Byte]): Byte = {
+    def takeAction(stockPrice: Double, marketState: List[Int]): Int = {
         // Evaluate the previous rule
         val updatedWealth = estimateWealth(stockPrice)
         if (currentAction != null){
